+++ sxg_rs/src/acme/mod.rs
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod jose;
+pub mod jws;
